% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_estabilidad_html.R
\name{analisis_estabilidad_html}
\alias{analisis_estabilidad_html}
\title{analisis_estabilidad_html}
\usage{
analisis_estabilidad_html(base, group_var, cont_vars, cat_train, name, direc)
}
\arguments{
\item{base}{Nombre de base de datos a analizar.}

\item{group_var}{Variable por la cual se prentende agrupar}

\item{cont_vars}{Nombre variables continuas}

\item{cat_train}{Nombres grupo train}

\item{name}{Nombre archivo a generar, este argumento debe ser entregado en formato:
"nombre.pdf"}

\item{direc}{Directorio donde se almacena el archivo a generar.}
}
\description{
Analisis de estabilidad
}
