% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asimetria_fisher.R
\name{asimetria_fisher}
\alias{asimetria_fisher}
\title{Coeficiente de Asimetria de Fisher}
\usage{
asimetria_fisher(x, interprete)
}
\arguments{
\item{x}{Nombre de la variable.}

\item{interprete}{Si desea que la funcion entrege interpretacion asignar
'SI' en este parametro, en caso contrario asignar 'NO'.}
}
\description{
Analiza la existencia de una distribucion asimetrica en una variable de tipo
continua.
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
