% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_performance_html.R
\name{binary_performance_html}
\alias{binary_performance_html}
\title{binary_performance_html}
\usage{
binary_performance_html(Y, pred, M, name, direc)
}
\arguments{
\item{Y}{Nombre de base de datos a analizar.}

\item{pred}{Variable por la cual se prentende agrupar}

\item{M}{Objeto con información adicional}

\item{name}{Nombre archivo a generar, este argumento debe ser entregado en formato:
"nombre.pdf"}

\item{direc}{Directorio donde se almacena el archivo a generar.}
}
\description{
Genera un informe estadistico que incorpora estadisticos descritivos
}
