% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_html.R
\name{descriptive_html}
\alias{descriptive_html}
\title{descriptive_html}
\usage{
descriptive_html(base, group_var, cont_vars, cat_vars, name, direc)
}
\arguments{
\item{base}{Nombre de base de datos a analizar.}

\item{group_var}{Variable por la cual se prentende agrupar}

\item{cont_vars}{Nombre variables continuas}

\item{cat_vars}{Nombres variables categoricas}

\item{name}{Nombre archivo a generar, este argumento debe ser entregado en formato:
"nombre.pdf"}

\item{direc}{Directorio donde se almacena el archivo a generar.}
}
\description{
Genera un informe estadistico que incorpora estadisticos descritivos
}
