% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_names.R
\name{edit_names}
\alias{edit_names}
\title{edit_names}
\usage{
edit_names(dataset, remains, new_format)
}
\arguments{
\item{dataset}{Nombre dataset}

\item{remains}{Vector de variables que no se editaran}

\item{new_format}{Terminacion del nuevo formato de nombres}
}
\description{
Cambia el formato de nombres de variables escogidas
}
