% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cont_html.R
\name{est_cont_html}
\alias{est_cont_html}
\title{est_cont_html}
\usage{
est_cont_html(base, var_cat, cont_var, name, direc)
}
\arguments{
\item{base}{Nombre de base de datos a analizar.}

\item{var_cat}{Nombre variable sobre la que se quiere agrupar}

\item{cont_var}{Nombres variables continuas}

\item{name}{Nombre archivo a generar, este argumento debe ser entregado en formato:
"nombre.pdf"}

\item{direc}{Directorio donde se almacena el archivo a generar.}
}
\description{
Genera un informe estadistico que incorpora estadisticos descritivos
}
\examples{
\dontrun{
data("albahaca")
informe.pdf(albahaca,"estadisticos.pdf","C:/Users/usuario/Desktop")}
}
