% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_binary_performance.R
\name{eval_binary_performance}
\alias{eval_binary_performance}
\title{eval_binary_performance}
\usage{
eval_binary_performance(Y, pred)
}
\arguments{
\item{Y}{Vector de valores reales}

\item{pred}{Vector de predicciones}

\item{text}{Titulo general de las graficas}
}
\description{
Calcula metricas de performance de un modelo de clasificacion binaria
}
