% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lift.R
\name{get_lift}
\alias{get_lift}
\title{get_lift}
\usage{
get_lift(pred, Y, groups = 100)
}
\arguments{
\item{pred}{Predicciones}

\item{Y}{Target}

\item{groups}{Cantidad de grupos}
}
\description{
Genera los valores dela lift por percentil
}
