% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{get_mode}
\usage{
get_mode(x, per = "BOTH")
}
\arguments{
\item{x}{Vector de factores}

\item{per}{Puede tomar los valores NO-> entregara solo la moda, SI-> entrega la proporcion de la categoria, BOTH-> entrega ambas}
}
\description{
Calcula la moda de una variable categorica
}
