% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_periods.R
\name{get_next_periods}
\alias{get_next_periods}
\title{Get next periods}
\usage{
get_next_periods(day, n)
}
\arguments{
\item{day}{Fecha}

\item{n}{Numero de periodos siguientes}
}
\description{
Obteniene los siguientes n periodos desde una fecha
}
