% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mes_to_period.R
\name{id_mes_to_period}
\alias{id_mes_to_period}
\title{id_mes_to_period}
\usage{
id_mes_to_period(idmes)
}
\arguments{
\item{idmes}{Fecha en formato \%Y\%m}
}
\description{
Transforma una fecha \%Y\%m a \%Y-\%m-01
}
