% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just_spread.R
\name{just_spread}
\alias{just_spread}
\title{just_spread}
\usage{
just_spread(var_num, data, ids, var_cat)
}
\arguments{
\item{var_num}{Nombre variable numerica}

\item{data}{Nombre dataset}

\item{ids}{Vector de nombres de ids}

\item{var_cat}{variable categrica}
}
\description{
Genera un spread del dataset
}
