% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_psi_test.R
\name{multiple_test_psi}
\alias{multiple_test_psi}
\title{multiple_test_psi}
\usage{
multiple_test_psi(var_cont, base, group_var, cat_train)
}
\arguments{
\item{var_cont}{nombre de la variable continua a analizar}

\item{base}{dataset}

\item{group_var}{variable que contiene los grupos}

\item{cat_train}{nombre grupo de train}
}
\description{
Calcula el indice psien multiples grupos de test
}
