% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_variable_psi.R
\name{multiple_variable_psi}
\alias{multiple_variable_psi}
\title{multiple_variable_psi}
\usage{
multiple_variable_psi(cont_vars, base, group_var, cat_train)
}
\arguments{
\item{cont_vars}{variables numericas}

\item{base}{dataset}

\item{group_var}{variable que contiene los grupos}

\item{cat_train}{nombre grupo de train}
}
\description{
Calcula para distintos grupos de test en distintan variables
}
