% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{Lista de graficos}

\item{file}{Archivo}

\item{cols}{Numero de graficos por file}

\item{layout}{layout}
}
\description{
Funcion de Cookbook for R, permite juntar multiples graficos en un output
}
