% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lift.R
\name{plot_lift}
\alias{plot_lift}
\title{get_lift}
\usage{
plot_lift(df, percentiles = c(5, 10, 15), test = "yes")
}
\arguments{
\item{df}{Dataset obtenido con get lift}

\item{percentiles}{Pecentiles a dibujar}

\item{test}{Indica si la lift graficada proviene de un dataset de test}
}
\description{
Genera los valores dela lift por percentil
}
