% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_index.R
\name{psi_index}
\alias{psi_index}
\title{psi_index}
\usage{
psi_index(cat_test, cat_train, base, var_cont, group_var)
}
\arguments{
\item{cat_test}{nombre grupo de test}

\item{cat_train}{nombre grupo de train}

\item{base}{dataset}

\item{var_cont}{nombre de la variable continua a analizar}

\item{group_var}{variable que contiene los grupos}
}
\description{
Calcula el indice psi
}
