% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_vector.R
\name{recode_vector}
\alias{recode_vector}
\title{recode_vector}
\usage{
recode_vector(vect)
}
\arguments{
\item{vect}{Vector}
}
\description{
Elimina caracteres extraños y espacios en blanco en un vector de strings
}
