% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_graph.R
\name{variable_graph}
\alias{variable_graph}
\title{Variable Graph}
\usage{
variable_graph(df, var)
}
\arguments{
\item{df}{Dataset}

\item{var}{Nombre variable a graficar}
}
\description{
Genera graficos de boxplot o barras según el tipo de variable
}
